A preliminary implementation of the `rayon` feature to speed up the
building of the `IndexedCrate` struct.

Results working on JSON from `aws-sdk-rust/aws-sdk-ec2`:

**Before:**

```
indexed_crate  fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ new         1.558 s       │ 1.675 s       │ 1.633 s       │ 1.631 s       │ 100     │ 100
```

**This commit:**

```
indexed_crate  fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ new         964.6 ms      │ 1.039 s       │ 997.9 ms      │ 998.7 ms      │ 100     │ 100
```

Closes #401

-----

I won't have time to work on this until this Friday (23rd August). So some thoughts in case someone wants to pick it up from here:

- Most of the gains come from parallelizing the `import_index`
- Parallelizing the `impl_index` properly is hard
- Some things benefit from being stored in a `BTreeMap` instead of a `HashMap` (or maybe pulling in a faster `Hash` impl)
- `.entry(key).or_default().push(value)` does not optimize as well as matching on the entry and inserting a `vec![value]` if vacant
- The `MapList` struct I introduced seems to hinder optimizations, so remove it where possible.
- Try to share more code between `rayon` and the default impl
- Switch the recursive functions for iterative ones as it shows up in the flamegraph