@obi1kenobi 2024-08-21
Good stuff so far! Excited to see where you take this next 🚀

All perf gains you make here get doubled in `cargo-semver-checks`, since it loads two crate versions' data per run. So you've already cut over a second from the `cargo-semver-checks` of that crate!

When you're happy with the perf gains and are getting ready to merge this, it would be good to move the conditionally-compiled bits into separate functions that we can conditionally include and call. I think this might make the code a bit easier to maintain in the long run.