@jalil-salame 2024-08-22
I tidied up the code and made sure it builds without rayon, I also profiled and most of the remaining time is spent on `VisibilityTracker::from_crate(crate_)` (roughly 60%-70% of the function time).

I'll refactor the `VisibilityTracker::from_crate(crate_)` function on Friday and see what perf gains I can achieve.

In the mean time could you check if the rayon stuff is easier to maintain now? The rayon and normal code is mixed together so that turning off the rayon feature affects the least amount of code.