@obi1kenobi 2024-08-22
Oh I almost forgot: we really want a CI step in the "run tests" job that ensures all tests pass both with and without the `rayon` feature.