@jalil-salame 2024-08-22
> Oh I almost forgot: we really want a CI step in the "run tests" job that ensures all tests pass both with and without the `rayon` feature.

Will do!

I wanted to get a Proof of Concept out ASAP so that I could ensure the things I'm doing align with your expectations/needs (and also to have something people could pick up and work with if I have to leave stuff halfway done).

I should have more time over the weekend to polish it up and get everything ready to merge c: