@jalil-salame 2024-08-22
I optimized the `VisibilityTracker::from_crate`. This shaves ~200ms off the numbers (no rayon 1.6s -> 1.4s) and with rayon (1s -> 0.8s).

There are probably some opportunities to parallelize it, but I'll work on that once I've had some sleep.