@obi1kenobi 2024-08-22
Nice find! 200ms before `rayon` is great! Thank you for all your hard work 🚀

Since the changes to `VisibilityTracker::from_crate` aren't related to `rayon`, I'd find them a lot easier to review if you made them in a separate PR. Right now, they are quite difficult to analyze in a PR that does everything all at once.

That PR would also benefit from having a robust benchmark that's part of the repo already.

In an ideal world, we'd have separate PRs like this:
- benchmark we use to analyze performance
- serial improvements to `VisibilityTracker::from_crate`
- `rayon` improvements to `VisibilityTracker::from_crate`
- `rayon` improvements to `IndexedCrate`

The fewer lines and less complexity there is in each PR, the easier the code will be to understand and make maintainable. In turn, that means we'll probably be able to find even more performance wins in the future!