@jalil-salame 2024-08-23
> Since the changes to `VisibilityTracker::from_crate` aren't related to `rayon`, I'd find them a lot easier to review if you made them in a separate PR. Right now, they are quite difficult to analyze in a PR that does everything all at once.

I don't know if you have seen this, but you can review commits individually: on the top left (right above the file tree in the review UI) you can change from `Changes from all commits` to changes from a specific commit. This adds a button to go to previous/next commit.

I can also split it up into multiple PRs, but because they all soft depend on each other I think this would be slightly easier to work on.

In preparation for either approach I split it up into multiple commits c: