@jalil-salame 2024-08-24
The profiler indicates that about ~80% of the time is now spent on `VisibilityTracker::from_crate`, and most of that time is spent on `HashMap` operations. Switching from `std::collections::HashMap` to `hashbrown::HashMap` or `rustc-hash::FxHashMap` might further improve performance.

An initial investigation on the improvements from switching to `rustc-hash::FxHashMap/Set` shows:

```console
$ cargo criterion --bench indexed_crate
IndexedCrate/new(aws-sdk-ec2)
                        time:   [1.3016 s 1.3060 s 1.3106 s]

$ cargo criterion --bench indexed_crate --features rustc-hash
IndexedCrate/new(aws-sdk-ec2)
                        time:   [1.1319 s 1.1352 s 1.1384 s]
                        change: [-13.469% -13.075% -12.695%] (p = 0.00 < 0.05)
                        Performance has improved.

$ cargo criterion --bench indexed_crate --features rayon
IndexedCrate/new(aws-sdk-ec2)
                        time:   [542.39 ms 543.88 ms 545.39 ms]
                        change: [-52.276% -52.090% -51.895%] (p = 0.00 < 0.05)
                        Performance has improved.

$ cargo criterion --bench indexed_crate --features rayon,rustc-hash
IndexedCrate/new(aws-sdk-ec2)
                        time:   [414.43 ms 415.55 ms 416.69 ms]
                        change: [-23.884% -23.596% -23.292%] (p = 0.00 < 0.05)
                        Performance has improved.
```