@obi1kenobi 2024-08-25
> I don't know if you have seen this, but you can review commits individually: on the top left (right above the file tree in the review UI) you can change from Changes from all commits to changes from a specific commit. This adds a button to go to previous/next commit.
>
> I can also split it up into multiple PRs, but because they all soft depend on each other I think this would be slightly easier to work on.

I have seen it, yes. Unfortunately, individual commits have other downsides relative to separate PRs.

The core issue is that I maintain (currently) 5 different branches of this project at the same time. To make that feasible and not a nightmare, I have a workflow tuned around squash-merging PRs and cherry-picking the resulting PR commit onto all the different branches.

We have to keep the different commits here separate, in case in the future we find one of them has an issue that requires it to be reverted. That means I can't squash-merge, which means I can't use my usual automation for maintaining this repo.

Sorry! I hope you understand now why I have to insist on separate PRs.