@jalil-salame 2024-08-26
> Sorry! I hope you understand now why I have to insist on separate PRs.

I understand, I assumed something like that would be the case, but I hoped it wouldn't be. Then I'll make PRs for each commit? Or should I group some of them together?