@obi1kenobi 2024-08-26
PRs for each commit would be great! Feel free to open the independent ones concurrently — or all of them concurrently, if you don't mind some more rebasing as we merge them.

Thanks for bearing with me here, I know this is an inconvenience and appreciate you being open to helping make my workflow easier!