@jalil-salame 2024-08-26
I split off the first three commits as they can be merged in parallel, the rest should cause merge conflicts so I'll create the PRs once they are merged.

My system seems to be noisy for perf measurements so expect results to be off by +/-2% (even when criterion says something is 2% faster, it's most likely noise).